+++ blocks/acf/example-acf/fields.php
@@ -6,17 +6,17 @@ if (!defined('ABSPATH')) {
 
 return array(
     'key'      => 'group_lonestar_example_acf',
-    'title'    => __('Lonestar Example ACF', 'lonestar-theme'),
+    'title'    => __('Lonestar Example ACF', 'lonestar'),
     'fields'   => array(
         array(
             'key'   => 'field_lonestar_example_acf_title',
-            'label' => __('Title', 'lonestar-theme'),
+            'label' => __('Title', 'lonestar'),
             'name'  => 'title',
             'type'  => 'text',
         ),
         array(
             'key'          => 'field_lonestar_example_acf_text',
-            'label'        => __('Text', 'lonestar-theme'),
+            'label'        => __('Text', 'lonestar'),
             'name'         => 'text',
             'type'         => 'textarea',
             'new_lines'    => 'wpautop',