+++ modules/module.acf-field-gravity-forms.php
@@ -23,7 +23,7 @@ function lonestar_module_register_acf_gravity_form_selector_field()
             public function __construct()
             {
                 $this->name = 'gravity_form_selector';
-                $this->label = __('Gravity Form Selector', 'lonestar-theme');
+                $this->label = __('Gravity Form Selector', 'lonestar');
                 $this->category = 'choice';
                 parent::__construct();
             }
@@ -48,7 +48,7 @@ function lonestar_module_register_acf_gravity_form_selector_field()
                     esc_attr($input_name),
                     esc_attr($input_class)
                 );
-                echo '<option value="">' . esc_html__('None', 'lonestar-theme') . '</option>';
+                echo '<option value="">' . esc_html__('None', 'lonestar') . '</option>';
 
                 foreach ($forms as $form) {
                     printf(